airflow/requirements.txt
apache-airflow==2.6.3
cassandra-driver==3.25.0
